design(#30): 채팅창 레이아웃 변경
